+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 Diretório dedicado a especificar o que foi atualizado do projeto a cada versão.
 
+# VERSÃO 0.6.0 - 06/06/2026
+- APLICADO ETL SILVER PARA AS TABELAS estoque_ingredientes e vendas_semanais
+
+
 # VERSÃO 0.5.0 - 28/05/2026
 - ARQUITETURA DE INGESTÃO FEITA
 - LOGGING E DOCKER FUNCIONANDO